Module Aset Rumah Tangga: - Can EDIT and DELETE.
Migration:
- Change data type from FLOAT => DOUBLE on table aset_rumah_tangga.
- Add column satuan VARCHAR[30] on table master_jenis_aset.

Seeder:
- Add column satuan on MasterJenisAsetSeeder.

+++ database/migrations/2016_04_24_082310_create_aset_rumah_tangga.php
@@ -18,11 +18,11 @@ class CreateAsetRumahTangga extends Migration
             
             $table->integer('id_master_jenis_aset');
             $table->integer('volume')->nullable();
-            $table->float('nilai_satuan')->nullable();
-            $table->float('nilai_total')->nullable();
+            $table->double('nilai_satuan')->nullable();
+            $table->double('nilai_total')->nullable();
             $table->integer('cara_perolehan')->nullable();
             $table->smallInteger('jenis_aset')->nullable();
-            $table->float('pendapatan_produktif')->nullable();
+            $table->double('pendapatan_produktif')->nullable();
 
             $table->timestamps();
             $table->softDeletes();
